+++ hello_world/android/app/src/main/kotlin/com/example/hello_world/MainActivity.kt
@@ -2,5 +2,4 @@ package com.example.hello_world
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
